+++ t4mon/orchestrator.py
@@ -11,13 +11,8 @@ import pandas as pd
 import types
 from multiprocessing import Pool
 
-# import threading
-
-from matplotlib import pylab as pylab  # isort:skip
-from matplotlib import pyplot as plt  # isort:skip
-
 from . import collector  # isort:skip
-from .logger import DEFAULT_LOGLEVEL, init_logger  # isort:skip
+from .logger import init_logger  # isort:skip
 from .gen_report import gen_report  # isort:skip
 from .df_tools import consolidate_data, reload_from_csv  # isort:skip
 
@@ -25,6 +20,7 @@ from .df_tools import consolidate_data, reload_from_csv  # isort:skip
 __all__ = ('Orchestrator')
 
 
+# Make the Orchestrator class pickable, required by Pool.map()
 def _pickle_method(method):
     func_name = method.im_func.__name__
     obj = method.im_self
@@ -47,13 +43,15 @@ copy_reg.pickle(types.MethodType, _pickle_method, _unpickle_method)
 
 class Orchestrator(object):
 
-    """ Object passed to jinja2 containing:
+    """
+    Class orchestrating the retrieval, local storage and report generation
+    It represents the object passed to jinja2 containing:
 
-        - graphs: dictionary of key, value -> {system-id, list of graphs}
+        - data: dataframe passed to get_graphs()
         - logs: dictionary of key, value -> {system-id, log entries}
         - date_time: Report generation date and time
-        - data: dataframe passed to get_graphs()
         - system: string containing current system-id being rendered
+
     """
 
     def __init__(self,
@@ -63,15 +61,12 @@ class Orchestrator(object):
                  settings_file=None,
                  safe=False,
                  **kwargs):
-        self.calculations_file = ''
+        self.data = pd.DataFrame()
         self.date_time = dt.date.strftime(dt.datetime.today(),
                                           "%d/%m/%Y %H:%M:%S")
-        self.date_tag = self.date_tag()
-        self.graphs = {}  # will be filled by calls from within jinja for loop
-        self.graphs_definition_file = ''
-        self.html_template = ''
         self.loglevel = loglevel
         self.logger = logger or init_logger(self.loglevel)
+        self.logs = {}
         self.noreports = noreports
         self.reports_written = []
         self.reports_folder = './reports'
@@ -79,49 +74,30 @@ class Orchestrator(object):
         self.store_folder = './store'
         self.safe = safe
         self.kwargs = kwargs
-        self._index = 0
-        self.data = pd.DataFrame()
-        self.logs = {}
         self.systems = [item for item in
                         collector.read_config(self.settings_file).sections()
                         if item not in ['GATEWAY', 'MISC']]
         self.check_files()
 
     def __str__(self):
-        return ('Orchestrator object created on {0} with loglevel {1}\n'
-                'graphs_definition_file: {2}\n'
-                'html_template: {3}\n'
-                'reports folder: {4}\n'
-                'store folder: {5}\n'
-                'data size: {6}\n'
-                'settings_file: {7}\n'
-                'calculations_file: {8}\n'
-                'mode: {9}'.format(
+        return ('Orchestrator object created on {} with loglevel {}\n'
+                'reports folder: {}\n'
+                'store folder: {}\n'
+                'data size: {}\n'
+                'settings_file: {}\n'
+                'mode: {}'.format(
                     self.date_time,
                     logging.getLevelName(self.logger.level),
-                    self.graphs_definition_file,
-                    self.html_template,
                     self.reports_folder,
                     self.store_folder,
                     self.data.shape,
                     self.settings_file,
-                    self.calculations_file,
                     'safe' if self.safe else 'fast'
                 ))
 
-    def __iter__(self):
-        return self
-
-    def next(self):  # Python 3: def __next__(self)
-        if self._index >= len(self.systems):
-            raise StopIteration
-        else:
-            self._index += 1
-            return self
-
     def get_absolute_path(self, filename=''):
         """
-        Return the absolute path if relative to the settings file location
+        Get the absolute path if relative to the settings file location
         """
         if os.path.isabs(filename):
             return filename
@@ -135,7 +111,7 @@ class Orchestrator(object):
         """
         Read the settings file and check if external config files (i.e.
         calculations_file, html_template) are defined and add them to the
-        Orchestrator objects
+        Orchestrator object as attributes
         """
         conf = collector.read_config(self.settings_file)
         self.logger.debug('Using settings file: %s', self.settings_file)
@@ -150,9 +126,31 @@ class Orchestrator(object):
                     self.get_absolute_path(conf.get('MISC', option))
                 )
 
+    def check_external_files_from_config(self):
+        """
+        Read the settings file and check if external config files (i.e.
+        calculations_file, html_template) are defined and exist in the
+        filesystem
+        """
+        conf = collector.read_config(self.settings_file)
+        self.logger.debug('Using settings file: %s', self.settings_file)
+        for option in ['calculations_file',
+                       'html_template',
+                       'graphs_definition_file',
+                       'reports_folder',
+                       'store_folder']:
+            value = self.get_absolute_path(conf.get('MISC', option))
+            if option in self.__dict__:  # update valid values
+                self.__setattr__(option, value)
+            elif not os.path.isfile(value):
+                self.logger.critical('%s NOT found: %s',
+                                     option,
+                                     value)
+                raise collector.ConfigReadError
+
     def check_folders(self):
         """
-        Runtime method checking if all destination folders are in place:
+        Check during runtime if all destination folders are in place:
 
         - reports output folder (self.reports_folder)
         - CSV and DAT store folder (self.store_folder)
@@ -174,7 +172,7 @@ class Orchestrator(object):
 
     def check_files(self):
         """
-        Runtime method checking if all required files exist and are readable:
+        Check during runtime if all required files exist and are readable:
 
         - settings file
         - calculations file (settings/MISC/calculations_file)
@@ -183,50 +181,40 @@ class Orchestrator(object):
         - reports output folder (self.reports_folder)
         - CSV and DAT store folder (self.store_folder)
 
-        Returns: Boolean (whether or not everything is in place)
+        Raises collector.ConfigReadError if not everything is in place
         """
         if not os.path.isfile(self.settings_file):
             self.logger.critical('Settings file not found: %s',
                                  self.settings_file)
             raise collector.ConfigReadError
         try:
-            self.get_external_files_from_config()
+            self.check_external_files_from_config()
         except (collector.ConfigReadError,
                 collector.ConfigParser.Error) as _exc:
             self.logger.exception(_exc)
             raise collector.ConfigReadError
-
         # Check that destination folders are in place
         self.check_folders()
 
-        for option in ['graphs_definition_file',
-                       'html_template',
-                       'calculations_file']:
-            try:
-                self_attribute = self.__getattribute__(option)
-            except AttributeError:
-                raise collector.ConfigReadError
-            self.logger.info('Checking %s: %s', option, self_attribute)
-            if not self_attribute or not os.path.isfile(self_attribute):
-                self.logger.critical('%s NOT found: %s',
-                                     option,
-                                     self_attribute)
-                raise collector.ConfigReadError
-
     def date_tag(self):
-        """ Converts self.date_time to the format '%Y%m%d_%H%M' """
+        """
+        Convert self.date_time to the filesystem friendly format '%Y%m%d_%H%M'
+        """
         current_date = dt.datetime.strptime(self.date_time,
                                             "%d/%m/%Y %H:%M:%S")
         return dt.date.strftime(current_date,
                                 "%Y%m%d_%H%M")
 
     def create_report(self, system=None, logger=None):
-        """ Method for creating a single report for a particular system """
+        """
+        Create a single report for a particular system
+        """
+        # TODO: allow creating a common report (comparison)
         if not system:
             raise AttributeError('Need a value for system!')
         logger = logger or init_logger(self.loglevel)
         report_name = '{0}/Report_{1}_{2}.html'.format(self.reports_folder,
-                                                       self.date_tag,
+                                                       self.date_tag(),
                                                        system)
         logger.debug('%s | Generating HTML report (%s)',
                      system,
@@ -236,40 +224,12 @@ class Orchestrator(object):
                                          system=system))
         return report_name
 
-    # def reports_generator(self):
-    #     """
-    #     Call jinja2 template, separately to safely store the logs
-    #     in case of error.
-    #     Doing this with threads throws many errors with Qt (when acting as
-    #     matplotlib backend out from main thread)
-    #     """
-    #     # Initialize default figure sizes and styling
-    #     pylab.rcParams['figure.figsize'] = 13, 10
-
-    #     plt.style.use('ggplot')
-
-    #     if self.safe:
-    #         for system in self.collector.systems:
-    #             self.create_report(system)
-    #     else:
-    #         threads = [
-    #             threading.Thread(target=self.create_report,
-    #                              kwargs={'system': system},
-    #                              name=system)
-    #             for system in self.collector.systems
-    #         ]
-    #         for thread_item in threads:
-    #             thread_item.daemon = True
-    #             thread_item.start()
-    #         for thread_item in threads:  # Assuming all take the same time
-    #             thread_item.join()
-
     def reports_generator(self):
         """
         Call jinja2 template, separately to safely store the logs
         in case of error.
-        Doing this with multiple processes to avoid problems with GC and
-        matplotlib backends under Windows environments.
+        Doing this with a multiprocessing pool to avoid problems with GC and
+        matplotlib backends under Windows environments
         """
         if self.safe:
             for system in self.systems:
@@ -286,13 +246,15 @@ class Orchestrator(object):
 
     def local_store(self, col):
         """
-        Makes a local copy of the current data in CSV and gzipped pickle
+        Make a local copy of the current data in CSV and gzipped pickle
         """
         self.logger.info('Making a local copy of data in store folder: ')
-        destfile = '{0}/data_{1}.pkl'.format(self.store_folder, self.date_tag)
+        destfile = '{0}/data_{1}.pkl'.format(self.store_folder,
+                                             self.date_tag())
         col.to_pickle(destfile, compress=True)
         self.logger.info('  -->  %s.gz', destfile)
-        destfile = '{0}/data_{1}.csv'.format(self.store_folder, self.date_tag)
+        destfile = '{0}/data_{1}.csv'.format(self.store_folder,
+                                             self.date_tag())
         col.data.to_csv(destfile)
         self.logger.info('  -->  %s', destfile)
 
@@ -304,14 +266,14 @@ class Orchestrator(object):
                     continue
                 with open('{0}/logs_{1}_{2}.txt'.format(self.store_folder,
                                                         system,
-                                                        self.date_tag),
+                                                        self.date_tag()),
                           'w') as logtxt:
                     logtxt.writelines(self.logs[system])
 
     def start(self):  # pragma: no cover
-        """ Main method, gets data and logs, store and render the HTML output
         """
-
+        Main method, get data and logs, store and render the HTML output
+        """
         # Open the connection and gather all data and logs
         _collector = collector.Collector(
             logger=self.logger,
@@ -344,7 +306,9 @@ class Orchestrator(object):
                                   pkl=True,
                                   plain=False,
                                   system=None):
-        """ Generate HTML files from data stored locally """
+        """
+        Generate HTML files from data stored locally
+        """
         # load the input file
         if not os.path.exists(data_file):
             self.logger.error('%s file %s cannot be found',
@@ -369,49 +333,9 @@ class Orchestrator(object):
         for system in self.systems:
             self.logs[system] = 'Log collection omitted for '\
                                 'locally generated reports at '\
-                                '{} for {}'.format(self.date_tag, system)
+                                '{} for {}'.format(self.date_tag(), system)
             self.logger.info(self.logs[system])
 
         # Create the reports
         self.reports_generator()
         self.logger.info('Done!')
-
-    # API CHANGE ON 0.12 MAKES THIS OBSOLETE
-    # def set_threaded_mode(self, safe=False):
-        # """ Change both orchestrator and collector mode (serial/threaded) """
-        # self.safe = self.collector.safe = safe
-        # self.logger.debug('Changed to %s mode',
-                          # 'safe' if safe else 'fast')
-
-    # def set_settings_file(self, settings_file=None):
-        # """ Change the settings file both for orchestrator and collector """
-        # self.settings_file = settings_file or collector.DEFAULT_SETTINGS_FILE
-        # self.collector.settings_file = self.settings_file
-        # self.collector.conf = collector.read_config(self.settings_file)
-
-    # def set_logger_level(self, loglevel=None):
-        # """ Change the loglevel for orchestrator and collector objects """
-        # if not loglevel:
-            # loglevel = DEFAULT_LOGLEVEL
-        # self.logger.setLevel(loglevel)
-        # self.collector.logger.setLevel(loglevel)
-
-
-def create_simple_report(self, logger=None):
-    """ Method for creating a single report for a particular system """
-    logger = logger or init_logger(self.loglevel)
-    logger.critical(self._index)
-    system = self.systems[self._index]
-    if not system:
-        raise AttributeError('Need a value for system!')
-    report_name = '{0}/Report_{1}_{2}.html'.format(self.reports_folder,
-                                                   self.date_tag,
-                                                   system)
-    logger.debug('%s | Generating HTML report (%s)',
-                 system,
-                 report_name)
-    with open(report_name, 'w') as output:
-        output.writelines(gen_report(container=self,
-                                     system=system,
-                                     logger=logger))
-    return report_name